xgitberg

+++ gitensite/apps/bookinfo/management/commands/load_repos.py
@@ -9,10 +9,14 @@ from gitenberg.util.catalog import BookMetadata, repo_list
 
 class Command(BaseCommand):
     help = "load ids and repo names from Raymond's text TSV file"
-    args = "rdf_library"
-    def handle(self, rdf_library, *args, **kwargs):
-        Book.objects.all().delete()
+    args = "rdf_library", "start"
+    def handle(self, rdf_library, start=0, *args, **kwargs):
+        start = int(start)
+        if start==0:
+            Book.objects.all().delete()
         for (pg_id, repo_name) in repo_list:
+            if int(pg_id)<start:
+                continue
             try:
                 (book,created) = Book.objects.get_or_create(book_id=int(pg_id), repo_name=repo_name)
                 metadata=BookMetadata(book,rdf_library=rdf_library, enrich=True)
